+++ packages/go_router/lib/src/router.dart
@@ -531,10 +531,18 @@ class GoRouter implements RouterConfig<RouteMatchList> {
       // verified by assert() during the initialization.
       return initialLocation!;
     }
-    final Uri platformDefaultUri =
-        Uri.parse(WidgetsBinding.instance.platformDispatcher.defaultRouteName);
-    final String platformDefault =
-        platformDefaultUri.path.isEmpty ? '/' : platformDefaultUri.path;
+    Uri platformDefaultUri = Uri.parse(
+      WidgetsBinding.instance.platformDispatcher.defaultRouteName,
+    );
+    if (platformDefaultUri.hasEmptyPath) {
+      // TODO(chunhtai): Clean up this once `RouteInformation.uri` is available
+      // in packages repo.
+      platformDefaultUri = Uri(
+        path: '/',
+        queryParameters: platformDefaultUri.queryParameters,
+      );
+    }
+    final String platformDefault = platformDefaultUri.toString();
     if (initialLocation == null) {
       return platformDefault;
     } else if (platformDefault == '/') {