+++ packages/in_app_purchase/in_app_purchase_storekit/lib/src/store_kit_wrappers/sk_payment_queue_wrapper.dart
@@ -41,7 +41,19 @@ class SKPaymentQueueWrapper {
   SKPaymentQueueDelegateWrapper? _paymentQueueDelegate;
   SKTransactionObserverWrapper? _observer;
 
-  /// Calls [`-[SKPaymentQueue transactions]`](https://developer.apple.com/documentation/storekit/skpaymentqueue/1506026-transactions?language=objc)
+  /// Calls [`[SKPaymentQueue storefront]`](https://developer.apple.com/documentation/storekit/skpaymentqueue/3182430-storefront?language=objc).
+  ///
+  /// Returns `null` if the user's device is below iOS 13.0 or macOS 10.15.
+  Future<SKStorefrontWrapper?> storefront() async {
+    final Map<String, dynamic>? storefrontMap = await channel
+        .invokeMapMethod<String, dynamic>('-[SKPaymentQueue storefront]');
+    if (storefrontMap == null) {
+      return null;
+    }
+    return SKStorefrontWrapper.fromJson(storefrontMap);
+  }
+
+  /// Calls [`-[SKPaymentQueue transactions]`](https://developer.apple.com/documentation/storekit/skpaymentqueue/1506026-transactions?language=objc).
   Future<List<SKPaymentTransactionWrapper>> transactions() async {
     return _getTransactionList((await channel
         .invokeListMethod<dynamic>('-[SKPaymentQueue transactions]'))!);