+++ packages/pigeon/example/app/macos/Runner/messages.g.m
@@ -16,6 +16,29 @@
 #error File requires ARC to be enabled.
 #endif
 
+static NSArray *wrapResult(id result, FlutterError *error) {
+  if (error) {
+    return @[
+      error.code ?: [NSNull null], error.message ?: [NSNull null], error.details ?: [NSNull null]
+    ];
+  }
+  return @[ result ?: [NSNull null] ];
+}
+
+static FlutterError *createConnectionError(NSString *channelName) {
+  return [FlutterError
+      errorWithCode:@"channel-error"
+            message:[NSString stringWithFormat:@"%@/%@/%@",
+                                               @"Unable to establish connection on channel: '",
+                                               channelName, @"'."]
+            details:@""];
+}
+
+static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
+  id result = array[key];
+  return (result == [NSNull null]) ? nil : result;
+}
+
 @implementation PGNCodeBox
 - (instancetype)initWithValue:(PGNCode)value {
   self = [super init];
@@ -25,19 +48,6 @@
   return self;
 }
 @end
-
-static NSArray *wrapResult(id result, FlutterError *error) {
-  if (error) {
-    return @[
-      error.code ?: [NSNull null], error.message ?: [NSNull null], error.details ?: [NSNull null]
-    ];
-  }
-  return @[ result ?: [NSNull null] ];
-}
-static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
-  id result = array[key];
-  return (result == [NSNull null]) ? nil : result;
-}
 
 @interface PGNMessageData ()
 + (PGNMessageData *)fromList:(NSArray *)list;
@@ -213,11 +223,12 @@ NSObject<FlutterMessageCodec> *PGNMessageFlutterApiGetCodec(void) {
 }
 - (void)flutterMethodAString:(nullable NSString *)arg_aString
                   completion:(void (^)(NSString *_Nullable, FlutterError *_Nullable))completion {
-  FlutterBasicMessageChannel *channel = [FlutterBasicMessageChannel
-      messageChannelWithName:
-          @"dev.flutter.pigeon.pigeon_example_package.MessageFlutterApi.flutterMethod"
-             binaryMessenger:self.binaryMessenger
-                       codec:PGNMessageFlutterApiGetCodec()];
+  NSString *channelName =
+      @"dev.flutter.pigeon.pigeon_example_package.MessageFlutterApi.flutterMethod";
+  FlutterBasicMessageChannel *channel =
+      [FlutterBasicMessageChannel messageChannelWithName:channelName
+                                         binaryMessenger:self.binaryMessenger
+                                                   codec:PGNMessageFlutterApiGetCodec()];
   [channel sendMessage:@[ arg_aString ?: [NSNull null] ]
                  reply:^(NSArray<id> *reply) {
                    if (reply != nil) {
@@ -230,10 +241,7 @@ NSObject<FlutterMessageCodec> *PGNMessageFlutterApiGetCodec(void) {
                        completion(output, nil);
                      }
                    } else {
-                     completion(nil, [FlutterError
-                                         errorWithCode:@"channel-error"
-                                               message:@"Unable to establish connection on channel."
-                                               details:@""]);
+                     completion(nil, createConnectionError(channelName));
                    }
                  }];
 }