+++ packages/pigeon/tool/shared/process_utils.dart
@@ -3,27 +3,19 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:io' show Process, stderr, stdout;
-
-Future<Process> _streamOutput(Future<Process> processFuture) async {
-  final Process process = await processFuture;
-  await Future.wait(<Future<Object?>>[
-    stdout.addStream(process.stdout),
-    stderr.addStream(process.stderr),
-  ]);
-  return process;
-}
+import 'dart:io' show Process, ProcessStartMode, stderr, stdout;
 
 Future<int> runProcess(String command, List<String> arguments,
     {String? workingDirectory,
     bool streamOutput = true,
     bool logFailure = false}) async {
-  final Future<Process> future = Process.start(
+  final Process process = await Process.start(
     command,
     arguments,
     workingDirectory: workingDirectory,
+    mode:
+        streamOutput ? ProcessStartMode.inheritStdio : ProcessStartMode.normal,
   );
-  final Process process = await (streamOutput ? _streamOutput(future) : future);
   final int exitCode = await process.exitCode;
   if (exitCode != 0 && logFailure) {
     // ignore: avoid_print