+++ packages/webview_flutter/webview_flutter/example/integration_test/webview_flutter_test.dart
@@ -537,7 +537,7 @@ Future<void> main() async {
       )));
       unawaited(controller.setOnScrollPositionChange(
           (ScrollPositionChange contentOffsetChange) {
-            recordedPosition = contentOffsetChange;
+        recordedPosition = contentOffsetChange;
       }));
 
       unawaited(controller.loadRequest(Uri.parse(