+++ packages/google_sign_in/google_sign_in_android/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 6.1.21
 
+* Updates `clearAuthCache` override to match base class declaration.
 * Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.
 
 ## 6.1.20